+++ skrub/__init__.py
@@ -56,7 +56,7 @@ from ._squashing_scaler import SquashingScaler
 from ._string_encoder import StringEncoder
 from ._table_vectorizer import Cleaner, TableVectorizer
 from ._tabular_pipeline import tabular_pipeline
-from ._text_encoder import TextEncoder
+from ._text_encoder import LLMEncoder, TextEncoder
 from ._to_categorical import ToCategorical
 from ._to_datetime import ToDatetime, to_datetime
 from ._to_float import ToFloat
@@ -84,7 +84,7 @@ __all__ = [
     "ToFloat",
     "ToCategorical",
     "TableVectorizer",
-    "TextEncoder",
+    "LLMEncoder",
     "StringEncoder",
     "Cleaner",
     "DropSimilar",
@@ -119,4 +119,5 @@ __all__ = [
     "config_context",
     "SessionEncoder",
     "core",
+    "TextEncoder",
 ]